Split embed components into separate file in new directory
